+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/impl/Accessor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2020, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -217,6 +217,8 @@ public abstract class Accessor {
         public abstract void exportSymbol(Object polyglotLanguageContext, String symbolName, Object value);
 
         public abstract Map<String, ? extends Object> getExportedSymbols();
+
+        public abstract Object getPolyglotBindingsObject();
 
         public abstract Object importSymbol(Object polyglotLanguageContext, Env env, String symbolName);